CONSOLE-3644: `app.jsx` -> `app.tsx`
