+++ frontend/public/locales/en/public.json
@@ -60,9 +60,6 @@
   "Instances": "Instances",
   "Skip to content": "Skip to content",
   "Modal": "Modal",
-  "Web console update is available": "Web console update is available",
-  "There has been an update to the web console. Ensure any changes have been saved and refresh your browser to access the latest version.": "There has been an update to the web console. Ensure any changes have been saved and refresh your browser to access the latest version.",
-  "Refresh web console": "Refresh web console",
   "Start build": "Start build",
   "Start last run": "Start last run",
   "BuildConfig details": "BuildConfig details",
@@ -1160,6 +1157,9 @@
   "Init containers": "Init containers",
   "CrashLoopBackOff": "CrashLoopBackOff",
   "Create Pod": "Create Pod",
+  "Web console update is available": "Web console update is available",
+  "There has been an update to the web console. Ensure any changes have been saved and refresh your browser to access the latest version.": "There has been an update to the web console. Ensure any changes have been saved and refresh your browser to access the latest version.",
+  "Refresh web console": "Refresh web console",
   "Service monitor selector": "Service monitor selector",
   "Promethesuses": "Promethesuses",
   "Quick create": "Quick create",
