+++ frontend/webpack.config.ts
@@ -62,7 +62,7 @@ const sharedPluginModulesTest = getVendorModuleRegExp(
 const config: Configuration = {
   entry: {
     main: [
-      './public/components/app.jsx',
+      './public/components/app.tsx',
       '/node_modules/@patternfly-5/patternfly/patternfly.scss',
       '/node_modules/@patternfly-5/patternfly/patternfly-addons.scss',
       '/node_modules/@patternfly-5/patternfly/patternfly-charts.scss',
